refactor!: rename "model" module to "function"
